subclass NetworkStateLiveData
