+++ app/src/main/java/de/xikolo/network/jobs/base/NetworkCode.kt
@@ -1,0 +1,5 @@
+package de.xikolo.network.jobs.base
+
+enum class NetworkCode {
+    STARTED, SUCCESS, ERROR, CANCEL, NO_NETWORK, NO_AUTH, MAINTENANCE, API_VERSION_EXPIRED, API_VERSION_DEPRECATED
+}
